requirements/base.txt
asgiref==3.5.2
Django==4.1.1
django-filter==22.1
django-rest-registration==0.7.2
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
PyJWT==2.5.0
pytz==2022.2.1
sqlparse==0.4.2
